New Log parser for commits
I've added a new log parser for commits which makes use of various features of git-log.
This fixes lots of bugs in the old, like that we have now the full message, always, in the database for the commit. I atm don't know any bugs in it, but it needs further testing, too.