porting #303680 by CorniI: New log parser.
Rewrote on several places, making the first 6.x version that just works.
This patch also include some minor changes.

CorniI:
The new parser now understands merges and supports both annotated and
unannotated tags. Branches are now supported so that if you add a
branch, all commits which are in that branch are adjusted, too. This
needs to be optimized further, I hope to get an API for that in the next
VCS-API so we don’t have to delete and re-add the commit... I can’t say
anything about performance in comparison to the old module, I never had
it working. For now, we have performance issues when importing mid-size
repos and when adding branches. I've an idea for the latter, though :)